Rotational power and Minecarts (#327)
This update introduces Vanderlin's rotational power method and the minecart system they have in place. Currently, there is only waterwheels to get power, but that might change.

I also updated some of the recipe books and included placeholder files and code for expanding to fluid dynamics and an autosmither, but those need to be discussed more.

Co-authored-by: BluBerry016 <50649185+unit0016@users.noreply.github.com>